refactor: move Cubos to core
