+++ engine/include/cubos/engine/collisions/plugin.hpp
@@ -13,7 +13,7 @@
 
 #pragma once
 
-#include <cubos/engine/cubos.hpp>
+#include <cubos/engine/prelude.hpp>
 
 namespace cubos::engine
 {
