+++ engine/include/cubos/engine/gizmos/gizmos.hpp
@@ -11,7 +11,7 @@
 
 #include <cubos/core/gl/render_device.hpp>
 
-#include <cubos/engine/cubos.hpp>
+#include <cubos/engine/prelude.hpp>
 
 namespace cubos::engine
 {
