+++ engine/include/cubos/engine/scene/plugin.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include <cubos/engine/cubos.hpp>
+#include <cubos/engine/prelude.hpp>
 #include <cubos/engine/scene/scene.hpp>
 
 namespace cubos::engine
