+++ engine/include/cubos/engine/settings/plugin.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include <cubos/engine/cubos.hpp>
+#include <cubos/engine/prelude.hpp>
 #include <cubos/engine/settings/settings.hpp>
 
 namespace cubos::engine
