+++ engine/include/cubos/engine/splitscreen/plugin.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include <cubos/engine/cubos.hpp>
+#include <cubos/engine/prelude.hpp>
 
 namespace cubos::engine
 {
