+++ tools/tesseratos/include/tesseratos/debug_camera/plugin.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include <cubos/engine/cubos.hpp>
+#include <cubos/engine/prelude.hpp>
 
 namespace tesseratos
 {
