+++ tools/tesseratos/include/tesseratos/world_inspector/plugin.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include <cubos/engine/cubos.hpp>
+#include <cubos/engine/prelude.hpp>
 
 namespace tesseratos
 {
